Disable more video tests on selected machines
Move the check to test_utils
Disable tests that try to use codecs that are unsupported on M60.
